requirements.txt
streamlit==1.33.0
transformers==4.35.2
sentence-transformers==2.2.2
accelerate==0.27.2
huggingface_hub==0.20.3
faiss-cpu
trafilatura
nltk
torch>=2.1.0
numpy==1.22.4
pandas
scipy==1.10.1
